LZMR-47 implemented locations service
