+++ Netbeans/SuperheroSightings/src/main/java/com/sg/superherosightings/service/LocationsService.java
@@ -1,0 +1,29 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.sg.superherosightings.service;
+
+import com.sg.superherosightings.dto.Location;
+import com.sg.superherosightings.dto.Sighting;
+import java.util.List;
+
+/**
+ *
+ * @author laptop
+ */
+public interface LocationsService {
+    
+    public Location addLocation(Location location);
+
+    public void deleteLocation(int locationId);
+
+    public void updateLocation(Location location);
+
+    public Location getLocationById(int locationId);
+
+    public List<Location> getAllLocations();
+
+    public List<Location> getAssociatedLocations(List<Sighting> sightings);
+}
